requirements.txt
rich>=13.7.0
redis>=5.0.0
psutil>=5.9.0
cryptography>=41.0.0
pyjwt>=2.8.0
